feat(migration) adding offline sqlx migration

+++ backend/migrations/20260405163546_fingerprints_schema.sql
@@ -1,7 +1,7 @@
 -- Add migration script here
 -- Fingerprints table (the big one)
 CREATE TABLE fingerprints (
-    hash INTEGER NOT NULL,
+    hash BIGINT NOT NULL,
     track_id UUID NOT NULL REFERENCES tracks(id) ON DELETE CASCADE,
     offset_ms INTEGER NOT NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
